Refactor DiagramApiController
