+++ src/ServiceControl.Monitoring/Infrastructure/Api/IEndpointMetricsApi.cs
@@ -1,0 +1,15 @@
+﻿namespace ServiceControl.Monitoring.Infrastructure.Api
+{
+    using ServiceControl.Monitoring.Http.Diagrams;
+
+    public interface IEndpointMetricsApi
+    {
+        MonitoredEndpoint[] GetAllEndpointsMetrics(int? history = null);
+
+        MonitoredEndpointDetails GetSingleEndpointMetrics(string endpointName, int? history = null);
+
+        void DeleteEndpointInstance(string endpointName, string instanceId);
+
+        int DisconnectedEndpointCount();
+    }
+}
