+++ src/ServiceControl.Monitoring/Infrastructure/WebApi/HostApplicationBuilderExtensions.cs
@@ -1,7 +1,7 @@
 namespace ServiceControl.Monitoring.Infrastructure.WebApi;
 
 using System.Reflection;
-using Infrastructure;
+using Api;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
@@ -9,6 +9,8 @@ public static class HostApplicationBuilderExtensions
 {
     public static void AddServiceControlMonitoringApi(this IHostApplicationBuilder hostBuilder)
     {
+        hostBuilder.Services.AddSingleton<IEndpointMetricsApi, EndpointMetricsApi>();
+
         var controllers = hostBuilder.Services.AddControllers(options =>
         {
             options.Filters.Add<XParticularVersionHttpHandler>();
